+++ CHANGELOG.md
@@ -5,6 +5,7 @@ development:
   - add `controller.sync-committee-message-delay` option
   - add `controller.sync-committee-aggregation-delay` option
   - submit attestations in batches, to speed up initial broadcast
+  - break down scheduler job start metrics by class
 
 1.2.2:
   - fix crash when no metrics configuration is supplied